rename instruct-lab plugin to attention-and-distributed-packing
