+++ .github/workflows/build-and-publish.yml
@@ -14,6 +14,7 @@ jobs:
           - "framework"
           - "accelerated-peft"
           - "fused-ops-and-kernels"
+          - "attention-and-distributed-packing"
 
     permissions:
       id-token: write  # IMPORTANT: this permission is mandatory for trusted publishing
