+++ plugins/attention-and-distributed-packing/.isort.cfg
@@ -8,6 +8,3 @@ import_heading_firstparty=First Party
 import_heading_localfolder=Local
 known_firstparty=
 known_localfolder=tuning
-
-# skip code imported from unsloth
-skip_glob=**/unsloth*/**
