+++ plugins/attention-and-distributed-packing/.pylintrc
