+++ sample-configurations/aadp-padding-free-sample-configuration.yaml
